+++ mage/src/main/java/mil/nga/giat/mage/map/GeocoderTask.java
@@ -10,22 +10,25 @@ import com.bricolsoftconsulting.geocoderplus.Geocoder;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 
 import java.io.IOException;
+import java.text.ParseException;
 import java.util.List;
 
 import mil.nga.giat.mage.sdk.connectivity.ConnectivityUtility;
+import mil.nga.mgrs.MGRS;
 
-public class GeocoderTask extends AsyncTask<String, Void, Address> {
+public class GeocoderTask extends AsyncTask<String, Void, MarkerOptions> {
 
 	private static final String LOG_NAME = GeocoderTask.class.getName();
 
 	private Context context;
+
 	// reference to map
 	private GoogleMap map;
+
 	// reference to search marker
 	private List<Marker> markers;
 
@@ -36,26 +39,43 @@ public class GeocoderTask extends AsyncTask<String, Void, Address> {
 	}
 
 	@Override
-	protected Address doInBackground(String... params) {
-		Address address = null;
+	protected MarkerOptions doInBackground(String... params) {
 		String searchString = params[0];
-		// Creating an instance of Geocoder class
-		Geocoder geocoder = new Geocoder();
+		MarkerOptions markerOptions = null;
 
-		try {
-			List<Address> addresses = geocoder.getFromLocationName(searchString, 1);
-			if (addresses != null && addresses.size() > 0) {
-				address = addresses.get(0);
+		if (MGRS.isMGRS(searchString)) {
+			try {
+				mil.nga.mgrs.wgs84.LatLng latLng = mil.nga.mgrs.wgs84.LatLng.parse(searchString);
+				markerOptions = new MarkerOptions()
+						.position(new LatLng(latLng.latitude, latLng.longitude))
+						.title("MGRS")
+						.snippet(searchString);
+			} catch (ParseException e) {
+				Log.e(LOG_NAME, "Problem parsing mgrs string", e);
 			}
-		} catch (IOException e) {
-			Log.e(LOG_NAME, "Problem executing search.");
+		} else {
+			// Creating an instance of Geocoder class
+			Geocoder geocoder = new Geocoder();
+
+			try {
+				List<Address> addresses = geocoder.getFromLocationName(searchString, 1);
+				if (addresses != null && addresses.size() > 0) {
+					Address address = addresses.get(0);
+					markerOptions = new MarkerOptions()
+						.position(new LatLng(address.getLatitude(), address.getLongitude()))
+						.title(searchString)
+						.snippet(address.getFormattedAddress());
+				}
+			} catch (IOException e) {
+				Log.e(LOG_NAME, "Problem executing search.");
+			}
 		}
 
-		return address;
+		return markerOptions;
 	}
 
 	@Override
-	protected void onPostExecute(Address result) {
+	protected void onPostExecute(MarkerOptions markerOptions) {
 		if (markers != null) {
 			for (Marker m : markers) {
 				m.remove();
@@ -63,36 +83,19 @@ public class GeocoderTask extends AsyncTask<String, Void, Address> {
 			markers.clear();
 		}
 
-		if (result == null) {
+		if (markerOptions == null) {
 			if(ConnectivityUtility.isOnline(context)) {
-				Toast.makeText(context, "Unknown location", Toast.LENGTH_SHORT).show();
+				Toast.makeText(context, "Unknown location", Toast.LENGTH_LONG).show();
 			} else {
 				Toast.makeText(context, "No connectivity, try again later.", Toast.LENGTH_LONG).show();
 			}
 		} else {
-			LatLng latLng = new LatLng(result.getLatitude(), result.getLongitude());
-			String addressText = result.getFormattedAddress();
-
-			if(addressText == null) {
-				addressText = "";
-			}
-			
-			MarkerOptions markerOptions = new MarkerOptions();
-			markerOptions.position(latLng);
-			markerOptions.title(addressText);
 			markers.add(map.addMarker(markerOptions));
 			for (Marker m : markers) {
 				m.showInfoWindow();
 			}
 
-			Log.d(LOG_NAME, "Search found: " + addressText);
-			
-			LatLng southWestlatLng = new LatLng(result.getViewPort().getSouthWest().getLatitude(), result.getViewPort().getSouthWest().getLongitude());
-			LatLng northEastlatLng = new LatLng(result.getViewPort().getNorthEast().getLatitude(), result.getViewPort().getNorthEast().getLongitude());
-			
-			map.animateCamera(CameraUpdateFactory.newLatLngBounds(new LatLngBounds(southWestlatLng, northEastlatLng), 10));
+			map.animateCamera(CameraUpdateFactory.newLatLngZoom(markerOptions.getPosition(), 10));
 		}
-
-		super.onPostExecute(result);
 	}
 }